Fix ex4 directory naming
